[render-ad-card] Дорабатывает реализацию отрисовки карточки объявления
